+++ index.js
@@ -1,4 +1,4 @@
 module.exports = {
   SubtitleParser: require('./lib/subtitle-parser'),
-  SubtitleStream: require('./lib/subtitle-stream-2')
+  SubtitleStream: require('./lib/subtitle-stream')
 }
