+++ lib/subtitle-stream.js
